Use 'latest' any-db instead of '*'.
If gesundheit is updated to use features (or fixes) in new any-db
version, existing `gesundheit` users will not get the updated
any-db version since any version can satisfy '*' thus breaking
the installation.

Further, `npm update gesundheit` won't fix it because '*' is
already satisfied with the buggy version (same issue).

I assumed that 'latest' is what is meant by specifying '*'.
That is, gesundheit is tightly coupled to any-db so there is no
point in specifying exact versions and both should be updated
together.